+++ nodes/RunNodeWithCredentialsX/v2/subWorkflowTemplate.json
@@ -1,0 +1,34 @@
+{
+  "nodes": [
+    {
+      "parameters": {},
+      "id": "dacc52e8-fef9-4cc3-9fca-623fa81b3c5f",
+      "name": "Start",
+      "type": "n8n-nodes-base.executeWorkflowTrigger",
+      "typeVersion": 1,
+      "position": [
+        720,
+        340
+      ]
+    }
+  ],
+	"settings":{
+		"saveManualExecutions":false,
+		"saveExecutionProgress":false,
+		"saveDataErrorExecution": "none",
+		"saveDataSuccessExecution": "none"
+	},
+  "connections": {
+    "Start": {
+      "main": [
+        [
+          {
+            "node": "",
+            "type": "main",
+            "index": 0
+          }
+        ]
+      ]
+    }
+  }
+}
