+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "n8n-nodes-run-node-with-credentials-x",
-  "version": "0.1.2",
+  "version": "0.2.0",
   "description": "Node to run any node with specified credentials ID.",
   "keywords": [
     "n8n-community-node-package"
